renamed audiosession to SoundPlayer and SoundPlayer to QuickPlay.
