+++ lib/src/plugins/base_plugin.dart
@@ -21,18 +21,18 @@ import 'dart:io';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
-import '../audio_session.dart';
 import '../codec.dart';
 import '../ios/ios_session_category.dart';
 import '../ios/ios_session_mode.dart';
 import '../playback_disposition.dart';
+import '../sound_player.dart';
 import '../track.dart';
 
 /// provides a set of common methods used by
 /// PluginInterfaces to talk to the underlying
 /// Platform specific plugin.
 abstract class BasePlugin {
-  final List<AudioSession> _slots = [];
+  final List<SoundPlayer> _slots = [];
 
   ///
   @protected
@@ -49,11 +49,11 @@ abstract class BasePlugin {
   }
 
   /// Over load this method to play audio.
-  Future<void> play(AudioSession session, Track track);
+  Future<void> play(SoundPlayer player, Track track);
 
   /// overload this method to handle callbacks from the underlying
   /// platform specific plugin
-  Future<dynamic> onMethodCallback(AudioSession session, MethodCall call);
+  Future<dynamic> onMethodCallback(SoundPlayer player, MethodCall call);
 
   Future<dynamic> _onMethodCallback(MethodCall call) {
     var slotNo = call.arguments['slotNo'] as int;
@@ -63,12 +63,12 @@ abstract class BasePlugin {
   }
 
   /// Invokes a method in the platform specific plugin for the
-  /// given [session]. The connector is a link either
-  /// a specific SoundRecorder or SoundPlayer instance.
+  /// given [player]. The connector is a link either
+  /// a specific SoundRecorder or QuickPlay instance.
   Future<dynamic> invokeMethod(
-      AudioSession session, String methodName, Map<String, dynamic> call) {
+      SoundPlayer player, String methodName, Map<String, dynamic> call) {
     /// allocate a slot for this call.
-    var slotNo = _findSlot(session);
+    var slotNo = _findSlot(player);
     call['slotNo'] = slotNo;
     return getChannel().invokeMethod<dynamic>(methodName, call);
   }
@@ -82,86 +82,86 @@ abstract class BasePlugin {
   /// with the platform specific plugin.
   /// To use a plugin you start by calling [register]
   /// and finish by calling [release].
-  void register(AudioSession session) {
+  void register(SoundPlayer player) {
     var inserted = false;
     for (var i = 0; i < _slots.length; ++i) {
       if (_slots[i] == null) {
-        _slots[i] = session;
+        _slots[i] = player;
         inserted = true;
         break;
       }
     }
     if (!inserted) {
-      _slots.add(session);
+      _slots.add(player);
     }
-    print('registered AudioSession to slot: ${_slots.length - 1}');
+    print('registered SoundPlayer to slot: ${_slots.length - 1}');
   }
 
   /// Releases the slot used by the connector.
   /// To use a plugin you start by calling [register]
   /// and finish by calling [release].
-  void release(AudioSession session) async {
-    await invokeMethod(session, 'releaseMediaPlayer', <String, dynamic>{});
+  void release(SoundPlayer player) async {
+    await invokeMethod(player, 'releaseMediaPlayer', <String, dynamic>{});
 
-    var slot = _findSlot(session);
+    var slot = _findSlot(player);
     if (slot != -1) {
       _slots[slot] = null;
     } else {
       throw AudioSessionNotRegisteredException(
-          'The AudioSession was not found when releasing the session.');
+          'The SoundPlayer was not found when releasing the player.');
     }
   }
 
   ///
-  void initialise(AudioSession session) async {
-    await invokeMethod(session, 'initializeMediaPlayer', <String, dynamic>{});
+  void initialise(SoundPlayer player) async {
+    await invokeMethod(player, 'initializeMediaPlayer', <String, dynamic>{});
   }
 
   ///
-  Future<void> stop(AudioSession session) async {
-    await invokeMethod(session, 'stopPlayer', <String, dynamic>{});
+  Future<void> stop(SoundPlayer player) async {
+    await invokeMethod(player, 'stopPlayer', <String, dynamic>{});
   }
 
   ///
-  Future<void> pause(AudioSession session) async {
-    await invokeMethod(session, 'pausePlayer', <String, dynamic>{});
+  Future<void> pause(SoundPlayer player) async {
+    await invokeMethod(player, 'pausePlayer', <String, dynamic>{});
   }
 
   ///
-  Future<void> resume(AudioSession session) async {
-    await invokeMethod(session, 'resumePlayer', <String, dynamic>{});
+  Future<void> resume(SoundPlayer player) async {
+    await invokeMethod(player, 'resumePlayer', <String, dynamic>{});
   }
 
   ///
-  Future<void> seekToPlayer(AudioSession session, Duration position) async {
-    await invokeMethod(session, 'seekToPlayer', <String, dynamic>{
+  Future<void> seekToPlayer(SoundPlayer player, Duration position) async {
+    await invokeMethod(player, 'seekToPlayer', <String, dynamic>{
       'sec': position.inMilliseconds,
     });
   }
 
   ///
-  Future<void> setVolume(AudioSession session, double volume) async {
+  Future<void> setVolume(SoundPlayer player, double volume) async {
     var indexedVolume = Platform.isIOS ? volume * 100 : volume;
     if (volume < 0.0 || volume > 1.0) {
       throw RangeError('Value of volume should be between 0.0 and 1.0.');
     }
-    await invokeMethod(session, 'setVolume', <String, dynamic>{
+    await invokeMethod(player, 'setVolume', <String, dynamic>{
       'volume': indexedVolume,
     });
   }
 
   ///
-  Future<bool> isSupported(AudioSession session, Codec codec) async {
-    var result = await invokeMethod(session, 'isDecoderSupported',
+  Future<bool> isSupported(SoundPlayer player, Codec codec) async {
+    var result = await invokeMethod(player, 'isDecoderSupported',
         <String, dynamic>{'codec': codec.index}) as bool;
     return result;
   }
 
   ///
-  Future<bool> iosSetCategory(AudioSession session, IOSSessionCategory category,
+  Future<bool> iosSetCategory(SoundPlayer player, IOSSessionCategory category,
       IOSSessionMode mode, int options) async {
     if (!Platform.isIOS) return false;
-    var r = await invokeMethod(session, 'iosSetCategory', <String, dynamic>{
+    var r = await invokeMethod(player, 'iosSetCategory', <String, dynamic>{
       'category': iosSessionCategory[category.index],
       'mode': iosSessionMode[mode.index],
       'options': options
@@ -171,16 +171,16 @@ abstract class BasePlugin {
   }
 
   ///
-  Future<bool> androidFocusRequest(AudioSession session, int focusGain) async {
+  Future<bool> androidFocusRequest(SoundPlayer player, int focusGain) async {
     if (!Platform.isAndroid) return false;
-    return await invokeMethod(session, 'androidAudioFocusRequest',
+    return await invokeMethod(player, 'androidAudioFocusRequest',
         <String, dynamic>{'focusGain': focusGain}) as bool;
   }
 
   ///
   Future<void> setSubscriptionDuration(
-      AudioSession session, Duration interval) async {
-    await invokeMethod(session, 'setSubscriptionDuration', <String, dynamic>{
+      SoundPlayer player, Duration interval) async {
+    await invokeMethod(player, 'setSubscriptionDuration', <String, dynamic>{
       /// we need to use milliseconds as if we use seconds we end
       /// up rounding down to zero.
       'sec': (interval.inMilliseconds).toDouble() / 1000,
@@ -189,17 +189,17 @@ abstract class BasePlugin {
 
   ///
   @protected
-  int _findSlot(AudioSession session) {
+  int _findSlot(SoundPlayer player) {
     var slot = -1;
     for (var i = 0; i < _slots.length; ++i) {
-      if (_slots[i] == session) {
+      if (_slots[i] == player) {
         slot = i;
         break;
       }
     }
     if (slot == -1) {
       throw AudioSessionNotRegisteredException(
-          'The AudioSession was not found.');
+          'The SoundPlayer was not found.');
     }
     return slot;
   }
@@ -207,17 +207,16 @@ abstract class BasePlugin {
   ///  The caller can manage his audio focus with this function
   /// Depending on your configuration this will either make
   /// this player the loudest stream or it will silence all other stream.
-  Future<void> requestAudioFocus(AudioSession session) async {
-    await invokeMethod(
-        session, 'setActive', <String, dynamic>{'enabled': true});
+  Future<void> requestAudioFocus(SoundPlayer player) async {
+    await invokeMethod(player, 'setActive', <String, dynamic>{'enabled': true});
   }
 
   ///  The caller can manage his audio focus with this function
   /// Depending on your configuration this will either make
   /// this player the loudest stream or it will silence all other stream.
-  Future<void> abandonAudioFocus(AudioSession session) async {
+  Future<void> abandonAudioFocus(SoundPlayer player) async {
     await invokeMethod(
-        session, 'setActive', <String, dynamic>{'enabled': false});
+        player, 'setActive', <String, dynamic>{'enabled': false});
   }
 
   /// Contrucsts a PlaybackDisposition from a json object.
