+++ lib/src/ui/slider.dart
