+++ lib/src/ui/slider_position.dart
